Simplify CSS
This removes a _bunch_ of now-unused CSS, and simplifies the entire thing. I'm
only setting what I actually need to set now, and I've consolidated that
configuration into a single file so that my CSS isn't so scattered throughout
the repo. Hopefully this should be easier to maintain in the long run.